added multiple indexing options

+++ src/DVN-web/src/edu/harvard/hmdc/vdcnet/web/servlet/VDCIndexServlet.java
@@ -9,8 +9,15 @@ package edu.harvard.hmdc.vdcnet.web.servlet;
 import edu.harvard.hmdc.vdcnet.admin.NetworkRoleServiceLocal;
 import edu.harvard.hmdc.vdcnet.admin.VDCUser;
 import edu.harvard.hmdc.vdcnet.index.IndexServiceLocal;
+import edu.harvard.hmdc.vdcnet.study.Study;
+import edu.harvard.hmdc.vdcnet.vdc.VDC;
+import edu.harvard.hmdc.vdcnet.vdc.VDCServiceLocal;
 import java.io.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.StringTokenizer;
 import javax.ejb.EJB;
+import javax.ejb.EJBException;
 
 import javax.servlet.*;
 import javax.servlet.http.*;
@@ -23,6 +30,7 @@ import javax.servlet.http.*;
 public class VDCIndexServlet extends HttpServlet {
     
     @EJB IndexServiceLocal indexService;
+    @EJB VDCServiceLocal vdcService;
     
     private boolean isNetworkAdmin(HttpServletRequest req) {
         VDCUser user = null;
@@ -48,13 +56,23 @@ public class VDCIndexServlet extends HttpServlet {
         
         if (isNetworkAdmin(req)) {
             beginPage(out);
-            out.println("<h3>Reindex confirmation</h3>");
-            out.println("To reindex, click on the button below.<body>");
-            out.print("<form method=POST><input value=Reindex type=submit /></form>");
+            out.println("<h3>Index Admin</h3>");
+            out.println("<form method=POST>");
+            out.println("To reindex all studies, click on the button below.<br/>");
+            out.print("<input name=index value=All type=submit />");
+            out.print("<hr>");
+            out.println("To index studies owned by a specific dataverse, input the dataverse id and click on the button below.<br/>");
+            out.print("<input name=\"vdcToIndex\" size=4>");
+            out.print("<input name=index value=Dataverse type=submit />");
+            out.print("<hr>");
+            out.println("To index arbitrary studies, input the study ids and click on the button below.<br/>");
+            out.print("<textarea name=\"studyIds\" rows=10></textarea><br/>");
+            out.print("<input name=index value=Studies type=submit />");
+            out.println("</form>");
             endPage(out);
             
         } else {
-            displayUnauthorizedPage(out);
+            displayMessage(out, "You are not authorized for this action.");
         }
     }
     
@@ -66,32 +84,80 @@ public class VDCIndexServlet extends HttpServlet {
     throws ServletException, IOException {
         res.setContentType("text/html;charset=UTF-8");
         PrintWriter out = res.getWriter();
+
+        if (isNetworkAdmin(req)) {        
+            String indexParam = req.getParameter("index");
         
-        if (isNetworkAdmin(req)) {
-            //first delete files
-            boolean deleteFailed = false;
-            File indexDir = new File("index-dir");
-            File[] files = indexDir.listFiles();
-            for (int i = 0; i < files.length; i++) {
-                if ( !files[i].delete() ) {
-                    deleteFailed = true;
+            if ("Dataverse".equals(indexParam)) {
+                Long vdcToIndex = null;
+                try {
+                    vdcToIndex = new Long(req.getParameter("vdcToIndex"));
+                    VDC vdc =  vdcService.findById(vdcToIndex);
+                    if (vdc != null) { 
+                        List studyIDList = new ArrayList();
+                        for (Study study :  vdc.getOwnedStudies() ) {
+                            studyIDList.add( study.getId() );
+                        }
+                        indexService.indexList( studyIDList );
+                        displayMessage (out,"Indexing succeeded.", "(for dataverse id = " + vdcToIndex + ")");
+                    } else {
+                        displayMessage (out, "Indexing failed.", "There is no dataverse with dvId = " + vdcToIndex);                    
+                    }
+                } catch (NumberFormatException nfe) {
+                    displayMessage (out, "Indexing failed.", "The dataverse id must be of type Long.");
+                } catch (Exception e) {
+                    e.printStackTrace();
+                    displayMessage (out, "Indexing failed.", "An unknown error occurred trying to index dataverse with id = " + vdcToIndex);
                 }
                 
-            }
-            
-            beginPage(out);
-            if (!deleteFailed) {
-                indexService.indexAll();
-                 out.println("<h3>Reindexing succeeded.</h3>");
+            } else if ("Studies".equals(indexParam)) {
+                String studyIds = req.getParameter("studyIds");
+                List studyIdList = new ArrayList();
+                String failedTokens = "";
+
+                StringTokenizer st = new StringTokenizer(studyIds, ",; \t\n\r\f");
+                 while (st.hasMoreTokens()) {
+                     String token = st.nextToken();
+                     try {
+                        studyIdList.add( new Long(token) );
+                     } catch (NumberFormatException nfe) {
+                        if (!failedTokens.equals("")) {
+                            failedTokens += ", ";
+                        }
+                        failedTokens += "\"" +token + "\"";
+                    }
+                 }
+                indexService.indexList( studyIdList );
+                if (!failedTokens.equals("")) {
+                    failedTokens = "(However, the following tokens were not of type Long: " + failedTokens + ")";
+                }
+                displayMessage (out,"Indexing succeeded.", failedTokens);
+
+            } else if ("All".equals(indexParam)) {
+                //first delete files
+                boolean deleteFailed = false;
+                File indexDir = new File("index-dir");
+                File[] files = indexDir.listFiles();
+                for (int i = 0; i < files.length; i++) {
+                    if ( !files[i].delete() ) {
+                        deleteFailed = true;
+                    }
+
+                }
+
+                if (!deleteFailed) {
+                    indexService.indexAll();
+                     displayMessage (out,"Reindexing succeeded.");
+                } else {
+                    displayMessage (out,"Reindexing failed.","There was a problem deleting the files. Please fix this manually, then try again.");
+                }
+
             } else {
-                out.println("<h3>Reindexing failed.</h3>");
-                out.println("There was a problem deleting the files. Please fix this manually, then try again.");
+                displayMessage (out, "You have selected an action that is not allowed.");
             }
-            endPage(out);
-            
         } else {
-            displayUnauthorizedPage(out);
-        }
+            displayMessage(out, "You are not authorized for this action.");
+        }        
     }
     
     private void beginPage(PrintWriter out) {
@@ -106,10 +172,17 @@ public class VDCIndexServlet extends HttpServlet {
         out.println("</html>");
         out.close();
     }
+
+    private void displayMessage(PrintWriter out, String title) {
+        displayMessage(out, title, null);
+    }
     
-    private void displayUnauthorizedPage(PrintWriter out) {
+    private void displayMessage(PrintWriter out, String title, String message) {
         beginPage(out);
-        out.println("<h3>You are not authorized for this action.</h3>");
+        out.println("<h3>" + title + "</h3>");
+        if (message != null && !message.trim().equals("") ) {
+            out.println(message);
+        }
         endPage(out);
     }
